refactor: remove excessive callback typedefs, simplify docs, move expandable code from discord-templates.h to discord.h
